backend/risk-prediction/requirements.txt
# Risk Prediction Service Dependencies

# FastAPI and web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.23

# ML and Transformers
torch>=2.0.0
transformers>=4.35.0
scipy>=1.11.0
numpy>=1.24.0
protobuf>=3.20.0
sacremoses>=0.0.53
sacremoses>=0.0.53

# Authentication
python-jose[cryptography]==3.3.0

# HTTP client for service communication
httpx>=0.25.2

# Utilities
python-dotenv>=1.0.0
